separate Header and FormHeader
